move function to util
